Introduce RelativeLevel
RelativeLevel is a complex leveling system that focus on the following
aspects:

 - give more weight to recent games than older games
 - favor players that play different opponents (vs always the same)
 - give more points when you beat a stronger player
 - win little points when you beat a weaker player

This should allow to keep the difference of levels lower,
competition and encourage players to play more.

To achieve this, I needed more info in the UpdateLevel function, so I
added the whole archive of both player, allowing this function to pick
the informations it needs from it.